tests for copy() moved to separate file
