+++ docs/api/frame/copy.rst
@@ -1,3 +1,4 @@
 
 .. xmethod:: datatable.Frame.copy
     :src: c/frame/py_frame.cc Frame::copy
+    :tests: tests/frame/test-copy.py
